+++ src/constants/api.js
@@ -2,4 +2,6 @@ export const LOGIN_URL = '/api/v1/auth/sign_in';
 export const LOGOUT_URL = 'api/v1/auth/sign_out';
 export const PROFILE_URL = '/api/v1/users/profile';
 export const STUDENT_URL = '/api/v1/students';
+export const PARENT_URL = '/api/v1/parents';
+export const ADDRESS_URL = '/api/v1/addresses';
 export const VALIDATE_TOKEN_URL = '/api/v1/auth/validate_token';
